chore: make CLI auth limits persist in ref-impl
